requirements.txt
ultralytics
filterpy
streamlit==1.25.0
streamlit-drawable-canvas==0.9.2
psycopg2-binary
python-dotenv
ffmpeg-python